remove unused func and sort imports
